add video right contianer to redux cycle
